refactored ids in FMLAdmin
FMLAdmin uses a lot of ids, it doesn't make sense to clutter with so many private methods when they're essentially all the same. Merged all but the screen_options one (which will be refactored later).

Also, got rid of page suffixes as they're only used in one place.

Finally, started refactor of the enqueuing of media-related styles and scripts

Signed-off-by: terry chay <tychay@php.net>