Refactored settings screen options with hidden columns version
It is better this way because it will be easier to add (default hidden) areas without too much code (no more custom javascript, ajax handlers, etc.).

There is this weird bug where we are not doing the admin_action_* actions on the settings page, so I added a hack to force that call. (I could have sworn this was working before)

Also removed old testing code form the settings template.

Signed-off-by: terry chay <tychay@php.net>