Refactor to deal with update_post_meta slash()ing everything
All data to update_post_meta is addslashes()'d. While a lot of this seems reduundant (wp_unslash()ing something right before we wp_slash() it), it is done this way to obey filter-input, escape-output. Maybe if WordPress did that, we wouldn't be in this mess

@see https://developer.wordpress.org/reference/functions/update_metadata/#source-code

Also, we now correctly store api data (if it has a \ in it.

Signed-off-by: terry chay <tychay@php.net>