deal with bug in esc_attr()
Apparently esc_attr doesn't escape carriage returns properly (or allows them in attributes).

in this case, if an alt is not provided and I'm using a caption template with returns in it, "wp_get_attachment_image" will make the alt be the caption which will inject the carriage returns and break the previous version of the extractor.

Signed-off-by: terry chay <tychay@php.net>