Option to use http flickr links instead of https
Only kicks in if 'image_prefer_http' is set and not on an ssl page.

Note that flickr redirects all http PAGES to https (not sure it's a great idea given the great firewall of China). So I only patch images (and the link to image), not page link injections.

Also added (future) support of attachment link default and removed some deprecated methods.

Signed-off-by: terry chay <tychay@php.net>